Created temperature sensor parser
This is used to read in the raw data from the sensor and pull
out a valid temperature. Testing to be done next.

The temperatures are accessed via the Pi's 1-Wire interface
which I have setup. Devices using 1-Wire have there outputs stored
within a system folder which is navigated to where a temp is
pulled out.